Fixed the recipe hyperlink color
I initially tried to change the color of the hyperlink using the class selectors, but it didn't work. After using type selectors, it worked and changed the color from default to black. This way, it'll look nicer and in the future I can be more specific if I want to when visited/clicked or not visited/clicked.